requirements.txt
Django==2.2.4
djangorestframework==3.10.2
psycopg2==2.8.3
numpy==1.17.0